Refactor modalize of MyAccount screen
