+++ src/ui/navigation/routeKeys.ts
@@ -14,7 +14,6 @@ export const deeplinkNavigatorScreen = 'DeeplinkNavigatorScreen' as const;
 export const mnemonicScreen = 'Mnemonic' as const;
 export const verifyMnemonicScreen = 'Verify Mnemonic' as const;
 export const createAccountScreen = 'Create Account' as const;
-export const balanceScreen = 'Account Balance' as const;
 
 export const dashboardScreen = 'Dashboard' as const;
 export const tipsScreen = 'Tips' as const;
@@ -26,8 +25,6 @@ export const motionDetailScreen = 'Motion' as const;
 export const notificationSettingsScreen = 'Notification' as const;
 export const accountsScreen = 'My Accounts' as const;
 export const myAccountScreen = 'My Account' as const;
-export const receiveFundScreen = 'ReceiveFund' as const;
-export const sendFundScreen = 'SendFund' as const;
 export const importAccountScreen = 'Import Account' as const;
 export const exportAccountWithJsonFileScreen = 'Export JSON' as const;
 export const bountiesScreen = 'Bounties' as const;
